updated for version 7.3.580 Problem:    Warning on 64 bit MS-Windows. Solution:   Add type cast. (Mike Williams)

+++ src/if_py_both.h
@@ -495,7 +495,7 @@ VimStrwidth(PyObject *self UNUSED, PyObject *args)
     if (!PyArg_ParseTuple(args, "s", &expr))
 	return NULL;
 
-    return PyLong_FromLong(mb_string2cells((char_u *)expr, STRLEN(expr)));
+    return PyLong_FromLong(mb_string2cells((char_u *)expr, (int)STRLEN(expr)));
 }
 
 /*
